short sequence + set default
added set default operator "s". syntax : s, <id>=<value> currently pitch + duration are implanted as ids.
added "short sequence" -- sequence can be in the form of (x/) and it will be read as (x/x).
this is not added for parallel, because that would mean the same note playing at the same time